Rename package to pytransform3d
