--- .gitlab-ci.yml
@@ -1,38 +1,0 @@
-# We assume that all dependencies are installed.
-
-stages:
-  - test
-
-ubuntu-python2-tests:
-  stage: test
-  script:
-    - nosetests -sv --with-coverage
-    - cd doc; make html
-  tags:
-    - ubuntu14.04,python2.7
-
-ubuntu-python3-tests:
-  stage: test
-  script:
-    - nosetests3 -sv --with-coverage
-    - cd doc; make html
-  tags:
-    - ubuntu14.04,python3.4
-
-windows-python2-tests:
-  stage: test
-  script:
-    - nosetests -sv --with-coverage
-    - cd doc
-    - make.bat html
-  tags:
-    - windows-server-2008,python2.7
-
-windows-python3-tests:
-  stage: test
-  script:
-    - nosetests -sv --with-coverage
-    - cd doc
-    - make.bat html
-  tags:
-    - windows-server-2008,python3.5
