+++ doc/source/conf.py
@@ -30,10 +30,10 @@ exclude_trees = ["_templates", "sphinxext"]
 source_suffix = '.rst'
 source_encoding = 'utf-8-sig'
 master_doc = 'index'
-project = u'pytransform'
+project = u'pytransform3d'
 copyright = u'2014, Alexander Fabisch'
-version = __import__("pytransform").__version__
-release = __import__("pytransform").__version__
+version = __import__("pytransform3d").__version__
+release = __import__("pytransform3d").__version__
 language = 'en'
 today_fmt = '%B %d, %Y'
 add_function_parentheses = True
