+++ examples/plot_collision_objects.py
@@ -7,7 +7,7 @@ This example shows how to load a URDF description of collision objects and
 display them.
 """
 import matplotlib.pyplot as plt
-from pytransform.urdf import UrdfTransformManager
+from pytransform3d.urdf import UrdfTransformManager
 
 
 URDF = """
