+++ examples/plot_transform.py
@@ -10,7 +10,7 @@ print(__doc__)
 
 
 import matplotlib.pyplot as plt
-from pytransform.transformations import plot_transform
+from pytransform3d.transformations import plot_transform
 
 
 plot_transform()
