+++ pytransform3d/camera.py
