+++ pytransform3d/editor.py
