+++ pytransform3d/rotations.py
