+++ pytransform3d/trajectories.py
