+++ setup.cfg
@@ -1,5 +1,5 @@
 [nosetests]
-cover-package=pytransform
+cover-package=pytransform3d
 with-doctest=true
 with-coverage=true
 cover-html=true
